% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_alert.R
\name{finish_alert}
\alias{finish_alert}
\title{Completion email alert}
\usage{
finish_alert(text = NULL, address_to = NULL, script_name = NULL,
  message_subject = NULL)
}
\arguments{
\item{text}{Optional text to include in email.}

\item{address_to}{email address where message will be sent. Default is that provided when running \code{set_up_script_alert()}.}

\item{script_name}{character string giving name of script.}

\item{message_subject}{subject of email message. Default is job name, start time, and server hostname.}
}
\value{
message subject
}
\description{
Completion email alert
}
