% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_alert.R
\name{set_up_script_alert}
\alias{set_up_script_alert}
\title{Set up script alert}
\usage{
set_up_script_alert()
}
\value{
\code{NULL}
}
\description{
Creates a file in the users home directory for future use.
}
